Update activity_serializer.py
I suggest changing the way we retrieve extra_context. In the function dumps, we are joining the parts by a comma, which may not be the best delimiter in the loads function because extra_contexts may have string fields that contain commas,and since the input of the loads function is a plain string, the list 'parts' in loads would be greater than 5 elements. 

if a user has an extra_context that contains a comma, pickle would throw an EOFError.